fix: 修复了Rust clippy警告
1. 为SkillRegistry实现Default trait，简化实例创建
2. 简化Option处理代码，使用map替代and_then+Some
3. 简化数据库查询字段默认值处理
4. 优化重试延迟计算，移除不必要的类型转换
5. 简化LLM响应解析中的filter_map使用
6. 合并重试条件判断代码，减少嵌套层级
7. 为枚举类型直接添加Default派生，移除手动实现
8. 添加clippy lint豁免注释以处理复杂类型和参数过多警告

+++ src-tauri/src/services/llm/gemini_adapter.rs
@@ -241,7 +241,7 @@ impl GeminiAdapter {
 
         for attempt in 0..=max_retries {
             if attempt > 0 {
-                let delay = Duration::from_millis(500 * 2u64.pow(attempt as u32 - 1));
+                let delay = Duration::from_millis(500 * 2u64.pow(attempt - 1));
                 log::warn!(
                     "请求重试, model={}, 第{}次重试, 延迟{}ms",
                     self.model,
